Added search for organization
