+++ dockstore-webservice/src/main/java/io/swagger/api/impl/ToolsApiServiceImpl.java
@@ -16,56 +16,43 @@
 
 package io.swagger.api.impl;
 
-import java.io.UnsupportedEncodingException;
-import java.net.MalformedURLException;
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.net.URLDecoder;
-import java.net.URLEncoder;
-import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.EnumSet;
-import java.util.List;
-import java.util.Objects;
-import java.util.Optional;
-import java.util.Set;
-import java.util.stream.Collectors;
-
-import javax.ws.rs.WebApplicationException;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.core.SecurityContext;
-
 import avro.shaded.com.google.common.base.Joiner;
 import com.google.common.base.Splitter;
-import com.google.common.collect.HashBasedTable;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Table;
-import com.google.gson.Gson;
 import io.dockstore.webservice.DockstoreWebserviceConfiguration;
 import io.dockstore.webservice.core.Entry;
 import io.dockstore.webservice.core.SourceFile;
-import io.dockstore.webservice.core.Tag;
 import io.dockstore.webservice.core.Tool;
 import io.dockstore.webservice.core.Version;
 import io.dockstore.webservice.core.Workflow;
-import io.dockstore.webservice.core.WorkflowVersion;
 import io.dockstore.webservice.helpers.EntryVersionHelper;
 import io.dockstore.webservice.jdbi.ToolDAO;
 import io.dockstore.webservice.jdbi.WorkflowDAO;
 import io.swagger.api.NotFoundException;
 import io.swagger.api.ToolsApiService;
-import io.swagger.model.ToolClass;
 import io.swagger.model.ToolDescriptor;
 import io.swagger.model.ToolDockerfile;
 import io.swagger.model.ToolVersion;
 import org.apache.commons.lang3.StringUtils;
-import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.apache.commons.lang3.tuple.Pair;
 import org.apache.http.HttpStatus;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+
+import javax.ws.rs.WebApplicationException;
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
+import javax.ws.rs.core.SecurityContext;
+import java.io.UnsupportedEncodingException;
+import java.net.MalformedURLException;
+import java.net.URI;
+import java.net.URISyntaxException;
+import java.net.URLDecoder;
+import java.nio.charset.StandardCharsets;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.Set;
 
 import static io.dockstore.webservice.core.SourceFile.FileType.CWL_TEST_JSON;
 import static io.dockstore.webservice.core.SourceFile.FileType.DOCKERFILE;
@@ -76,7 +63,7 @@ import static io.dockstore.webservice.core.SourceFile.FileType.WDL_TEST_JSON;
 public class ToolsApiServiceImpl extends ToolsApiService {
 
     public static final int DEFAULT_PAGE_SIZE = 1000;
-    private static final Logger LOG = LoggerFactory.getLogger(ToolsApiServiceImpl.class);
+
     private static ToolDAO toolDAO = null;
     private static WorkflowDAO workflowDAO = null;
     private static DockstoreWebserviceConfiguration config = null;
@@ -84,224 +71,12 @@ public class ToolsApiServiceImpl extends ToolsApiService {
     public static void setToolDAO(ToolDAO toolDAO) {
         ToolsApiServiceImpl.toolDAO = toolDAO;
     }
-
-    public static void setConfig(DockstoreWebserviceConfiguration config) {
-        ToolsApiServiceImpl.config = config;
-    }
-
     public static void setWorkflowDAO(WorkflowDAO workflowDAO) {
         ToolsApiServiceImpl.workflowDAO = workflowDAO;
     }
 
-    /**
-     * Convert our Tool object to a standard Tool format
-     *
-     * @param container our data object
-     * @return standardised data object
-     */
-    private static Pair<io.swagger.model.Tool, Table<String, SourceFile.FileType, Object>> convertContainer2Tool(Entry container) {
-        Table<String, SourceFile.FileType, Object> fileTable = HashBasedTable.create();
-        String globalId;
-        // TODO: properly pass this information
-        String newID;
-        try {
-            // construct escaped ID
-            if (container instanceof Tool) {
-                newID = ((Tool)container).getToolPath();
-            } else if (container instanceof Workflow) {
-                newID = "#workflow/" + ((Workflow)container).getPath();
-            } else {
-                LOG.error("Could not construct URL for our container with id: " + container.getId());
-                return null;
-            }
-
-            String escapedID = URLEncoder.encode(newID, StandardCharsets.UTF_8.displayName());
-            URI uri = new URI(config.getScheme(), null, config.getHostname(), Integer.parseInt(config.getPort()), "/api/ga4gh/v1/tools/",
-                    null, null);
-            globalId = uri.toString() + escapedID;
-        } catch (URISyntaxException | UnsupportedEncodingException e) {
-            LOG.error("Could not construct URL for our container with id: " + container.getId());
-            return null;
-        }
-        // TODO: hook this up to a type field in our DB?
-        ToolClass type = container instanceof Tool ? ToolClassesApiServiceImpl.getCommandLineToolClass()
-                : ToolClassesApiServiceImpl.getWorkflowClass();
-
-        io.swagger.model.Tool tool = new io.swagger.model.Tool();
-        if (container.getAuthor() == null) {
-            tool.setAuthor("Unknown author");
-        } else {
-            tool.setAuthor(container.getAuthor());
-        }
-
-        tool.setDescription(container.getDescription());
-        tool.setMetaVersion(container.getLastUpdated() != null ? container.getLastUpdated().toString() : new Date(0).toString());
-        tool.setToolclass(type);
-        tool.setId(newID);
-        tool.setUrl(globalId);
-
-        // tool specific
-        if (container instanceof Tool) {
-            Tool inputTool = (Tool)container;
-            tool.setToolname(inputTool.getToolname());
-            tool.setOrganization(inputTool.getNamespace());
-            tool.setToolname(inputTool.getName());
-        }
-        // workflow specific
-        if (container instanceof Workflow) {
-            Workflow inputTool = (Workflow)container;
-            tool.setToolname(inputTool.getPath());
-            tool.setOrganization(inputTool.getOrganization());
-            tool.setToolname(inputTool.getWorkflowName());
-        }
-
-        // TODO: contains has no counterpart in our DB
-        // setup versions as well
-        Set inputVersions;
-        if (container instanceof Tool) {
-            inputVersions = ((Tool)container).getTags();
-        } else {
-            inputVersions = ((Workflow)container).getWorkflowVersions();
-        }
-
-        // handle verified information
-        tool.setVerified(((Set<Version>)inputVersions).stream().anyMatch(Version::isVerified));
-        final List<String> collect = ((Set<Version>)inputVersions).stream().filter(Version::isVerified).map(Version::getVerifiedSource)
-                .collect(Collectors.toList());
-        Gson gson = new Gson();
-        tool.setVerifiedSource(gson.toJson(collect));
-
-        for (Version inputVersion : (Set<Version>)inputVersions) {
-
-            // tags with no names make no sense here
-            // also hide hidden tags
-            if (inputVersion.getName() == null || inputVersion.isHidden()) {
-                continue;
-            }
-            if (inputVersion instanceof Tag && ((Tag)inputVersion).getImageId() == null) {
-                continue;
-            }
-
-            ToolVersion version = new ToolVersion();
-            // version id
-            String globalVersionId;
-            try {
-                globalVersionId = globalId + "/versions/" + URLEncoder.encode(inputVersion.getName(), StandardCharsets.UTF_8.displayName());
-            } catch (UnsupportedEncodingException e) {
-                LOG.error("Could not construct URL for our container with id: " + container.getId());
-                return null;
-            }
-            version.setUrl(globalVersionId);
-
-            version.setId(tool.getId() + ":" + inputVersion.getName());
-
-            version.setName(inputVersion.getName());
-
-            version.setVerified(inputVersion.isVerified());
-            version.setVerifiedSource(inputVersion.getVerifiedSource());
-            version.setDockerfile(false);
-
-            String urlBuilt;
-            final String githubPrefix = "git@github.com:";
-            final String bitbucketPrefix = "git@bitbucket.org:";
-            if (container.getGitUrl().startsWith(githubPrefix)) {
-                urlBuilt = extractHTTPPrefix(container.getGitUrl(), inputVersion.getReference(), githubPrefix,
-                        "https://raw.githubusercontent.com/");
-            } else if (container.getGitUrl().startsWith(bitbucketPrefix)) {
-                urlBuilt = extractHTTPPrefix(container.getGitUrl(), inputVersion.getReference(), bitbucketPrefix, "https://bitbucket.org/");
-            } else {
-                LOG.error("Found a git url neither from bitbucket or github " + container.getGitUrl());
-                urlBuilt = null;
-            }
-
-            final Set<SourceFile> sourceFiles = inputVersion.getSourceFiles();
-            for (SourceFile file : sourceFiles) {
-                if (inputVersion instanceof Tag) {
-                    switch (file.getType()) {
-                    case DOCKERFILE:
-                        ToolDockerfile dockerfile = new ToolDockerfile();
-                        dockerfile.setDockerfile(file.getContent());
-                        dockerfile.setUrl(urlBuilt + ((Tag)inputVersion).getDockerfilePath());
-                        version.setDockerfile(true);
-                        fileTable.put(inputVersion.getName(), DOCKERFILE, dockerfile);
-                        break;
-                    case DOCKSTORE_CWL:
-                        version.addDescriptorTypeItem(ToolVersion.DescriptorTypeEnum.CWL);
-                        fileTable.put(inputVersion.getName(), DOCKSTORE_CWL,
-                                buildSourceFile(urlBuilt + ((Tag)inputVersion).getCwlPath(), file));
-                        break;
-                    case DOCKSTORE_WDL:
-                        version.addDescriptorTypeItem(ToolVersion.DescriptorTypeEnum.CWL);
-                        fileTable.put(inputVersion.getName(), DOCKSTORE_WDL,
-                                buildSourceFile(urlBuilt + ((Tag)inputVersion).getWdlPath(), file));
-                        break;
-                    }
-                } else if (inputVersion instanceof WorkflowVersion) {
-                    switch (file.getType()) {
-                    case DOCKSTORE_CWL:
-                        version.addDescriptorTypeItem(ToolVersion.DescriptorTypeEnum.CWL);
-                        fileTable.put(inputVersion.getName(), DOCKSTORE_CWL,
-                                buildSourceFile(urlBuilt + ((WorkflowVersion)inputVersion).getWorkflowPath(), file));
-                        break;
-                    case DOCKSTORE_WDL:
-                        version.addDescriptorTypeItem(ToolVersion.DescriptorTypeEnum.CWL);
-                        fileTable.put(inputVersion.getName(), DOCKSTORE_WDL,
-                                buildSourceFile(urlBuilt + ((WorkflowVersion)inputVersion).getWorkflowPath(), file));
-                        break;
-                    }
-                }
-            }
-            if (container instanceof Tool) {
-                version.setImage(((Tool)container).getPath() + ":" + inputVersion.getName());
-            }
-            // ensure that descriptor is non-null before adding to list
-            if (!version.getDescriptorType().isEmpty()) {
-                // do some clean-up
-                version.setMetaVersion(
-                        String.valueOf(inputVersion.getLastModified() != null ? inputVersion.getLastModified() : new Date(0)));
-                final List<ToolVersion.DescriptorTypeEnum> descriptorType = version.getDescriptorType();
-                if (!descriptorType.isEmpty()) {
-                    EnumSet<ToolVersion.DescriptorTypeEnum> set = EnumSet.copyOf(descriptorType);
-                    version.setDescriptorType(Lists.newArrayList(set));
-                }
-                tool.getVersions().add(version);
-            }
-
-        }
-        return new ImmutablePair<>(tool, fileTable);
-    }
-
-    /**
-     * Build a descriptor and attach it to a version
-     *
-     * @param url  url to set for the descriptor
-     * @param file a file with content for the descriptor
-     */
-    private static ToolDescriptor buildSourceFile(String url, SourceFile file) {
-        ToolDescriptor wdlDescriptor = new ToolDescriptor();
-        if (file.getType() == DOCKSTORE_CWL) {
-            wdlDescriptor.setType(ToolDescriptor.TypeEnum.CWL);
-        } else if (file.getType() == DOCKSTORE_WDL) {
-            wdlDescriptor.setType(ToolDescriptor.TypeEnum.WDL);
-        }
-        wdlDescriptor.setDescriptor(file.getContent());
-        wdlDescriptor.setUrl(url);
-        return wdlDescriptor;
-    }
-
-    /**
-     * @param gitUrl       The git formatted url for the repo
-     * @param reference    the git tag or branch
-     * @param githubPrefix the prefix for the git formatted url to strip out
-     * @param builtPrefix  the prefix to use to start the extracted prefix
-     * @return the prefix to access these files
-     */
-    private static String extractHTTPPrefix(String gitUrl, String reference, String githubPrefix, String builtPrefix) {
-        StringBuilder urlBuilder = new StringBuilder();
-        urlBuilder.append(builtPrefix);
-        final String substring = gitUrl.substring(githubPrefix.length(), gitUrl.lastIndexOf(".git"));
-        urlBuilder.append(substring).append('/').append(reference);
-        return urlBuilder.toString();
+    public static void setConfig(DockstoreWebserviceConfiguration config) {
+        ToolsApiServiceImpl.config = config;
     }
 
     @Override
@@ -326,7 +101,7 @@ public class ToolsApiServiceImpl extends ToolsApiService {
             // check whether this is registered
             response = Response.status(Response.Status.UNAUTHORIZED).build();
         } else {
-            io.swagger.model.Tool tool = convertContainer2Tool(container).getLeft();
+            io.swagger.model.Tool tool = ToolsImplCommon.convertContainer2Tool(container, config).getLeft();
             assert (tool != null);
             // filter out other versions if we're narrowing to a specific version
             if (version != null) {
@@ -453,8 +228,7 @@ public class ToolsApiServiceImpl extends ToolsApiService {
 
         List<io.swagger.model.Tool> results = new ArrayList<>();
         for (Entry c : all) {
-            if (c instanceof Workflow && (registryId != null || registry != null || organization != null || name != null
-                    || toolname != null)) {
+            if (c instanceof Workflow && (registryId != null || registry != null || organization != null || name != null || toolname != null)) {
                 continue;
             }
 
@@ -498,7 +272,7 @@ public class ToolsApiServiceImpl extends ToolsApiService {
                 }
             }
             // if passing, for each container that matches the criteria, convert to standardised format and return
-            io.swagger.model.Tool tool = convertContainer2Tool(c).getLeft();
+            io.swagger.model.Tool tool = ToolsImplCommon.convertContainer2Tool(c, config).getLeft();
             if (tool != null) {
                 results.add(tool);
             }
@@ -578,7 +352,7 @@ public class ToolsApiServiceImpl extends ToolsApiService {
             return Response.status(Response.Status.UNAUTHORIZED).build();
         }
 
-        final Pair<io.swagger.model.Tool, Table<String, SourceFile.FileType, Object>> toolTablePair = convertContainer2Tool(entry);
+        final Pair<io.swagger.model.Tool, Table<String, SourceFile.FileType, Object>> toolTablePair = ToolsImplCommon.convertContainer2Tool(entry, config);
 
         String finalVersionId = versionId;
         if (toolTablePair == null || toolTablePair.getKey().getVersions() == null) {
